+++ src/StringLiteralGenerator.Roslyn3/StringLiteralGenerator.Roslyn3.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<IsPackable>false</IsPackable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.11.0" PrivateAssets="all" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<Compile Include="../$(RootNamespace)/*.cs" Exclude="../$(RootNamespace)/Utf8StringLiteralGenerator.cs" />
+	</ItemGroup>
+</Project>
