+++ src/StringLiteralGenerator/StringLiteralGenerator.csproj
@@ -1,32 +1,36 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>preview</LangVersion>
-    <Nullable>enable</Nullable>
-    <IsRoslynComponent>true</IsRoslynComponent>
-    <IncludeBuildOutput>false</IncludeBuildOutput>
+	<PropertyGroup>
+		<DevelopmentDependency>true</DevelopmentDependency>
+		<PackageId>StringLiteralGenerator</PackageId>
+		<Authors>Nobuyuki Iwanaga</Authors>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<RepositoryUrl>https://github.com/ufcpp/StringLiteralGenerator</RepositoryUrl>
+		<RepositoryType>git</RepositoryType>
+		<PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+		<Description>A C# Source Generator for optimizing UTF-8 binaries.</Description>
+	</PropertyGroup>
 
-    <PackageId>StringLiteralGenerator</PackageId>
-    <Authors>Nobuyuki Iwanaga</Authors>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <RepositoryUrl>https://github.com/ufcpp/StringLiteralGenerator</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <Description>A C# Source Generator for optimizing UTF-8 binaries.</Description>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.0.1" PrivateAssets="all" />
+	</ItemGroup>
 
-    <TargetsForTfmSpecificContentInPackage>$(TargetsForTfmSpecificContentInPackage);_AddAnalyzersToOutput</TargetsForTfmSpecificContentInPackage>
-  </PropertyGroup>
+	<ItemGroup>
+		<!-- Package the generator in the analyzer directory of the nuget package -->
+		<None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/roslyn4.0/cs" Visible="false" />
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.0.0-4.final" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.2" PrivateAssets="all" />
-  </ItemGroup>
+		<!-- Package the props file -->
+		<Content Include="build\$(RootNamespace).targets" Pack="true" PackagePath="build" />
+	</ItemGroup>
 
-  <Target Name="_AddAnalyzersToOutput">
-    <ItemGroup>
-      <TfmSpecificPackageFile Include="$(OutputPath)\StringLiteralGenerator.dll" PackagePath="analyzers/dotnet/cs" />
-    </ItemGroup>
-  </Target>
+	<Target Name="_AddAnalyzersToOutput" BeforeTargets="_GetPackageFiles">
+		<MSBuild Projects="../$(RootNamespace).Roslyn3/$(RootNamespace).Roslyn3.csproj" Targets="GetTargetPath">
+			<Output ItemName="Roslyn3Assembly" TaskParameter="TargetOutputs" />
+		</MSBuild>
+
+		<ItemGroup>
+			<None Include="%(Roslyn3Assembly.Identity)" Pack="true" PackagePath="analyzers/dotnet/roslyn3.11/cs" Visible="false" />
+		</ItemGroup>
+	</Target>
 
 </Project>
