+++ src/StringLiteralGenerator/Utf8StringLiteralGenerator.Emitter.cs
@@ -1,30 +1,7 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Text;
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-using System.Text;
-
-namespace StringLiteralGenerator;
+﻿namespace StringLiteralGenerator;
 
 public partial class Utf8StringLiteralGenerator
 {
-    private static void Emit(SourceProductionContext context, ImmutableArray<Utf8LiteralMethod> methods)
-    {
-        var buffer = new StringBuilder();
-
-        var group = methods.GroupBy(x => x.Type, x => x.Method);
-
-        foreach (var g in group)
-        {
-            var containingType = g.Key;
-            var generatedSource = Generate(containingType, g, buffer);
-            var filename = GetFilename(containingType, buffer);
-            context.AddSource(filename, SourceText.From(generatedSource, Encoding.UTF8));
-        }
-    }
-
     private static string GetFilename(TypeInfo type, StringBuilder buffer)
     {
         buffer.Clear();
